audit: ByteOrderMask pin on uint32_t for LP64 safety
ByteOrderMask is a packed 16-lane field descriptor consumed by
SwapBytesBom; the kbom* literals (kbomSwapShort=0x40000000,
kbomSwapLong=0xC0000000, struct-specific masks like
kbomActnf=0xC0000000, etc.) all assume a 32-bit container.
Bare `unsigned long` is 4 bytes on x86 and Win64 LLP64, but 8 on
LP64 (Linux/Mac x64) — the typedef would silently widen and
SwapBytesBom would walk an extra 16 phantom 2-bit lanes off the
high half of the mask, treating uninitialised bits as swap directives.

Pin to uint32_t in kauai/src/utilint.h. Add #include <cstdint>
to the same header so it compiles standalone where its consumers
don't already pull cstdint via chunk.h.

Runtime-only typedef; no on-disk format change.

+++ kauai/src/utilint.h
@@ -16,6 +16,8 @@
 #ifndef UTILINT_H
 #define UTILINT_H
 
+#include <cstdint>
+
 /****************************************
     Scalar constants
 ****************************************/
@@ -310,8 +312,12 @@ inline long LwMul(long lw1, long lw2)
     Byte Swapping
 ****************************************/
 
-// byte order mask
-typedef ulong ByteOrderMask;
+// byte order mask: a packed array of 16 2-bit lane descriptors used by
+// SwapBytesBom to byteswap a struct field-by-field. Pinned to 32 bits via
+// uint32_t so the kbom* literals (e.g. 0xC0000000) keep the same bit layout
+// on every architecture; bare `unsigned long` would silently widen to 64 bits
+// on LP64 (Linux/Mac x64) and reinterpret the high half as extra fields.
+typedef uint32_t ByteOrderMask;
 
 void SwapBytesBom(void *pv, ByteOrderMask bom);
 void SwapBytesRgsw(void *psw, long csw);
